requirements.txt
python-telegram-bot=13.12
requests=2.32.3
python-dotenv=1.0.1